refactor: move some constants & methods to new systemctl_mqtt/_dbus.py
